Fix: Remove code clone by extracting shared test methods into AbstractAutoCompleterTest - resolves #59
